requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
alembic==1.13.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
asteroid-filterbanks==0.4.0
attrs==24.2.0
audioread==3.0.1
audiosegment==0.23.0
av==11.0.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.8.2
boto3==1.35.5
botocore==1.35.5
brave-search==0.1.8
build==1.2.1
cachetools==5.5.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cohere==5.8.1
coloredlogs==15.0.1
colorlog==6.8.2
contourpy==1.2.1
crewai==0.51.1
crewai-tools==0.8.3
ctranslate2==4.3.1
cycler==0.12.1
dataclasses-json==0.6.7
decorator==5.1.1
Deprecated==1.2.14
deprecation==2.1.0
diart==0.9.1
distro==1.9.0
docker==7.1.0
docopt==0.6.2
docstring_parser==0.16
docx2txt==0.8
duckduckgo_search==6.2.11
einops==0.8.0
embedchain==0.1.121
fastapi==0.112.2
fastavro==1.9.5
faster-whisper==1.0.0
ffmpeg-python==0.2.0
filelock==3.15.4
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SocketIO==5.3.6
flatbuffers==24.3.25
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
future==1.0.0
google-api-core==2.19.1
google-auth==2.34.0
google-cloud-aiplatform==1.63.0
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-cloud-resource-manager==1.12.5
google-cloud-storage==2.18.2
google-crc32c==1.5.0
google-resumable-media==2.7.2
googleapis-common-protos==1.63.2
gptcache==0.1.44
groq==0.9.0
grpc-google-iam-v1==0.13.1
grpcio==1.66.0
grpcio-status==1.62.3
grpcio-tools==1.62.3
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.25.2
httpx-sse==0.4.0
huggingface-hub==0.24.6
humanfriendly==10.0
hyperframe==6.0.1
HyperPyYAML==1.2.2
idna==3.8
importlib_metadata==8.0.0
importlib_resources==6.4.4
iniconfig==2.0.0
instructor==1.3.3
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.4.2
jmespath==1.0.1
joblib==1.4.2
json_repair==0.25.3
jsonpatch==1.33
jsonpointer==3.0.0
jsonref==1.1.0
julius==0.2.7
kiwisolver==1.4.5
kubernetes==30.1.0
lancedb==0.5.7
langchain==0.2.14
langchain-cohere==0.1.9
langchain-community==0.2.12
langchain-core==0.2.34
langchain-experimental==0.0.64
langchain-openai==0.1.22
langchain-text-splitters==0.2.2
langsmith==0.1.104
lazy_loader==0.4
librosa==0.10.2.post1
lightning==2.4.0
lightning-utilities==0.11.6
llvmlite==0.43.0
Mako==1.3.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.22.0
matplotlib==3.5.3
mdurl==0.1.2
mem0ai==0.0.20
mmh3==4.1.0
monotonic==1.6
more-itertools==10.4.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
mypy-extensions==1.0.0
networkx==3.3
nltk==3.9.1
nodeenv==1.9.1
numba==0.60.0
numpy==1.26.4
oauthlib==3.2.2
omegaconf==2.3.0
onnxruntime==1.19.0
openai-whisper @ git+https://github.com/openai/whisper.git@ba3f3cd54b0e5b8ce1ab3de13e32122d0d5f98ab
opentelemetry-api==1.26.0
opentelemetry-exporter-otlp-proto-common==1.26.0
opentelemetry-exporter-otlp-proto-grpc==1.26.0
opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-instrumentation==0.47b0
opentelemetry-instrumentation-asgi==0.47b0
opentelemetry-instrumentation-fastapi==0.47b0
opentelemetry-proto==1.26.0
opentelemetry-sdk==1.26.0
opentelemetry-semantic-conventions==0.47b0
opentelemetry-util-http==0.47b0
optuna==3.6.1
orjson==3.10.7
outcome==1.3.0.post0
overrides==7.7.0
packaging==24.1
pandas==2.2.2
parameterized==0.9.0
pillow==10.4.0
platformdirs==4.2.2
pluggy==1.5.0
pooch==1.8.2
portalocker==2.10.1
posthog==3.5.2
primePy==1.3
primp==0.6.1
proto-plus==1.24.0
protobuf==4.25.4
pulsar-client==3.5.0
py==1.11.0
pyannote.audio==3.1.1
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow==17.0.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pylance==0.9.18
pyparsing==3.1.2
pypdf==4.3.1
PyPika==0.48.9
pyproject_hooks==1.1.0
pyright==1.1.377
pysbd==0.3.4
PySocks==1.7.1
pytest==8.3.2
pytest-asyncio==0.23.8
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
pytorch-lightning==2.4.0
pytorch-metric-learning==2.6.0
pytube==15.0.0
pytz==2024.1
PyYAML==6.0.2
qdrant-client==1.11.0
ratelimiter==1.2.0.post0
regex==2023.12.25
requests==2.32.3
requests-oauthlib==2.0.0
retry==0.9.2
rich==13.7.1
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
Rx==3.2.0
s3transfer==0.10.2
safetensors==0.4.4
schema==0.7.7
scikit-learn==1.5.1
scipy==1.14.1
selenium==4.23.1
semver==3.0.2
sentencepiece==0.2.0
setuptools==72.2.0
shapely==2.0.6
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.0
soundfile==0.12.1
soupsieve==2.6
soxr==0.4.0
speechbrain==1.0.0
SpeechRecognition==3.10.4
SQLAlchemy==2.0.32
starlette==0.38.2
sympy==1.13.2
tabulate==0.9.0
tavily-python==0.4.0
tenacity==8.5.0
tensorboardX==2.6.2.2
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers==0.15.2
torch==2.4.0
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.4.0
torchmetrics==1.4.1
torchvision==0.19.0
tqdm==4.66.5
transformers==4.39.3
trio==0.26.2
trio-websocket==0.11.1
typer==0.12.5
types-requests==2.32.0.20240712
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.6
uvloop==0.20.0
watchfiles==0.23.0
webrtcvad==2.0.10
websocket-client==1.8.0
websocket-server==0.6.4
websockets==13.0
Werkzeug==3.0.4
wheel==0.44.0
whisper-live==0.0.11
wrapt==1.16.0
wsproto==1.2.0
yarl==1.9.4
zipp==3.20.0
